Fixing issues
1) Modify the `FreeRTOSConfig.h` and adding missing macros like (ISRs
   wrappers)

2) Fixing a bug in the `stm32f103_startup.cpp` which is wrong `.bss`
   initalization

3) Adding a test example to the hardware tests
