+++ dev/main.cpp
@@ -13,9 +13,6 @@
 #include "mcal/Gpio.h"
 #include "mcal/Rcc.h"
 #include "utils/Logger.h"
-#include "freeRTOS/FreeRTOS.h"
-#include "freeRTOS//task.h"
-
 
 using namespace stm32::registers::rcc;
 using namespace stm32::dev::mcal::pin;
@@ -24,47 +21,6 @@ using namespace stm32::dev::mcal::rcc;
 using namespace stm32::utils::logger;
 using namespace stm32::type;
 
-void T1_Handler(void* pvParameters);
-
 int main() {
-    TaskHandle_t T1Handle = NULL;
-    BaseType_t  xReturned;
-    Rcc::Init();
-    Gpio::Init();
-
-    Rcc::Enable(Peripheral::kUSART1);
-    Rcc::Enable(Peripheral::kIOPC);
-    
-    Rcc::InitSysClock();
-    Rcc::SetExternalClock(kHseCrystal);
-
-    xReturned = xTaskCreate(T1_Handler, "Task1", 100, NULL, 0, &T1Handle);
-    if (xReturned == pdPASS) {
-        Logger::Info("Task1 created successfully");
-    } else {
-        Logger::Error("Task1 creation failed");
-    }
-
-    /* Start the scheduler. */
-    vTaskStartScheduler();
     while (1) {}
-}
-
-void T1_Handler(void* pvParameters) {
-    (void)pvParameters;  //  Avoids unused variable warning
-    Pin pc13(kPortC, kPin13, PinMode::kOutputPushPull_10MHz);
-    Gpio::Set(pc13);
-    uint8_t toggleFlag = 1;
-
-    for (;;) {
-        Logger::Info("Task1 running...");
-        if (toggleFlag) {
-            Gpio::SetPinValue(pc13, kLow);
-            toggleFlag = 0;
-        } else {
-            Gpio::SetPinValue(pc13, kHigh);
-            toggleFlag = 1;
-        }
-        vTaskDelay(pdMS_TO_TICKS(500));  // Task delay for 500ms
-    }
 }
